+++ lib/src/main/java/io/ably/lib/realtime/ChannelEvent.java
@@ -4,13 +4,12 @@ package io.ably.lib.realtime;
  * Channel event
  */
 public enum ChannelEvent {
-    initialized,
-    attaching,
-    attached,
-    detaching,
-    detached,
-    failed,
-    suspended,
-    update
-
+	initialized,
+	attaching,
+	attached,
+	detaching,
+	detached,
+	failed,
+	suspended,
+	update
 }
