+++ lib/src/main/java/io/ably/lib/realtime/ConnectionEvent.java
@@ -4,13 +4,13 @@ package io.ably.lib.realtime;
  * Connection event
  */
 public enum ConnectionEvent {
-    initialized,
-    connecting,
-    connected,
-    disconnected,
-    suspended,
-    closing,
-    closed,
-    failed,
-    update
+	initialized,
+	connecting,
+	connected,
+	disconnected,
+	suspended,
+	closing,
+	closed,
+	failed,
+	update
 }
