+++ lib/src/main/java/io/ably/lib/transport/NetworkConnectivity.java
@@ -13,14 +13,21 @@ public abstract class NetworkConnectivity {
 	}
 
 	public void addListener(NetworkConnectivityListener listener) {
+		boolean wasEmpty = listeners.isEmpty();
 		listeners.add(listener);
+		if(wasEmpty) {
+			onNonempty();
+		}
 	}
 
 	public void removeListener(NetworkConnectivityListener listener) {
 		listeners.remove(listener);
+		if(listeners.isEmpty()) {
+			onEmpty();
+		}
 	}
 
-	protected static void notifyNetworkAvailable() {
+	protected void notifyNetworkAvailable() {
 		for(NetworkConnectivityListener listener: listeners) {
 			listener.onNetworkAvailable();
 		}
@@ -32,7 +39,26 @@ public abstract class NetworkConnectivity {
 		}
 	}
 
-	protected static Set<NetworkConnectivityListener> listeners = new HashSet<NetworkConnectivityListener>();
+	protected boolean isEmpty() {
+		return listeners.isEmpty();
+	}
+
+	protected void onEmpty() {}
+
+	protected void onNonempty() {}
+
+	protected Set<NetworkConnectivityListener> listeners = new HashSet<NetworkConnectivityListener>();
 
 	public static class DefaultNetworkConnectivity extends NetworkConnectivity {}
+
+	public static class DelegatedNetworkConnectivity extends NetworkConnectivity implements NetworkConnectivityListener {
+		@Override
+		public void onNetworkAvailable() {
+			notifyNetworkAvailable();
+		}
+		@Override
+		public void onNetworkUnavailable(ErrorInfo reason) {
+			notifyNetworkUnavailable(reason);
+		}
+	}
 }
