--- lib/src/main/java/io/ably/lib/types/Function.java
@@ -1,9 +1,0 @@
-package io.ably.lib.types;
-
-public interface Function<Arg, Result> {
-    public Result call(Arg arg);
-
-    public static interface Binary<Arg1, Arg2, Result> {
-        public Result call(Arg1 arg1, Arg2 arg2);
-    }
-}
