+++ lib/src/main/java/io/ably/lib/util/CurrentThreadExecutor.java
@@ -2,13 +2,11 @@ package io.ably.lib.util;
 
 import java.util.concurrent.Executor;
 
-
-
 public class CurrentThreadExecutor implements Executor {
-    public static CurrentThreadExecutor INSTANCE = new CurrentThreadExecutor();
+	public static CurrentThreadExecutor INSTANCE = new CurrentThreadExecutor();
 
-    @Override
-    public void execute(Runnable runnable) {
-        runnable.run();
-    }
+	@Override
+	public void execute(Runnable runnable) {
+		runnable.run();
+	}
 }
