+++ lib/src/main/java/io/ably/lib/util/StringUtils.java
@@ -5,13 +5,13 @@ import io.ably.lib.http.HttpCore;
 
 public class StringUtils {
 	public static Serialisation.FromJsonElement<String> fromJsonElement = new Serialisation.FromJsonElement<String>() {
-        @Override
-        public String fromJsonElement(JsonElement e) {
-            return e.getAsJsonPrimitive().getAsString();
-        }
-    };
+		@Override
+		public String fromJsonElement(JsonElement e) {
+			return e.getAsJsonPrimitive().getAsString();
+		}
+	};
 
-    public static HttpCore.ResponseHandler<String> httpResponseHandler = new Serialisation.HttpResponseHandler<String>(String.class, fromJsonElement);
+	public static HttpCore.ResponseHandler<String> httpResponseHandler = new Serialisation.HttpResponseHandler<String>(String.class, fromJsonElement);
 
-    public static HttpCore.BodyHandler<String> httpBodyHandler = new Serialisation.HttpBodyHandler<String>(String[].class, fromJsonElement);
+	public static HttpCore.BodyHandler<String> httpBodyHandler = new Serialisation.HttpBodyHandler<String>(String[].class, fromJsonElement);
 }
